requirements.txt
mkdocs-material
mkdocs-macros-plugin
pillow
cairosvg
